refactor: Collapse `VideoImporterMac` and `VideoImporterLinux` into `VideoImporter`
Preparatory work for adding cross-platform video support following the `PlatformImage` model in `ImageImporter`.
